requirements.txt
flask==3.0.0
pyyaml==6.0.1
easydict==1.11
torch==2.1.0
numpy==1.24.4
pandas==2.0.3
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
matplotlib==3.7.3
tqdm==4.66.1
loguru==0.7.2
torchvision==0.16.0
thop
tabulate==0.9.0
ultralytics==8.0.120
git+https://github.com/openai/CLIP.git
imgaug==0.4.0
labelme==5.3.1
pycocotools==2.0.7